backend/requirements.txt
asgiref==3.9.1
attrs==25.3.0
audioop-lts==0.2.2
autobahn==24.4.2
Automat==25.4.16
cachetools==5.5.2
certifi==2025.8.3
cffi==1.17.1
channels==4.3.1
channels_redis==4.3.0
charset-normalizer==3.4.3
comtypes==1.4.11
constantly==23.10.4
cryptography==45.0.6
daphne==4.2.1
Django==5.2.5
django-cors-headers==4.7.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
dnspython==2.7.0
dotenv==0.9.9
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
httplib2==0.22.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
msgpack==1.1.1
oauthlib==3.3.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyAudio==0.2.14
pycparser==2.22
pydub==0.25.1
PyJWT==2.10.1
pymongo==4.14.1
pyOpenSSL==25.1.0
pyparsing==3.2.3
pypiwin32==223
python-dotenv==1.1.1
pyttsx3==2.99
pywin32==311
redis==6.4.0
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
service-identity==24.2.0
setuptools==80.9.0
SpeechRecognition==3.14.3
sqlparse==0.5.3
standard-aifc==3.13.0
standard-chunk==3.13.0
Twisted==25.5.0
txaio==25.6.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
zope.interface==7.2
